XYZ curve

+++ SPH Sim/compactCell.hpp
@@ -1,6 +1,6 @@
 #pragma once
 struct CompactCell
 {
-	uint16_t particle;
+	size_t particle;
 	int cell;
 };
